Remove unnecessary Role enum from core/types Move PlanCriteria enum to core/enum Update docs
